pyproject.toml
[tool.poetry]
name = "foodrecipeapi"
version = "0.1.0"
description = ""
authors = ["Edward Turner <edward.turner@pyscale.net>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic-sqlalchemy = "^0.0.9"
graphene-pydantic = "^0.3.0"
fastapi = "^0.78.0"
graphene-sqlalchemy = "3.0.0b1"
graphql-relay = "^3.2.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.37"}
psycopg2 = "^2.9.3"
psycopg = "^3.0.14"
uvicorn = "^0.17.6"
starlette-exporter = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-postgresql = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
